Move load module into separated part.
