+++ core/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod datamodel;
 pub mod eval;
+pub mod load;
 pub mod repl;
